+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Full documentation for rocFFT is available at [rocfft.readthedocs.io](https://ro
   manual kernels.
 
 ### Optimizations
+- Improved performance of the standalone out-of-place transpose kernel
 - Optimized 1D length 40000 C2C case.
 - Enabled radix-7 for size 336.
 - New radix-11 and radix-13 kernels; used in length 11 and 13 (and some of their multiples) transforms.